requirements.txt
attrs==21.4.0
autoflake==1.4
bcrypt==3.2.0
black==21.9b0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
ciscoconfparse==1.5.46
click==8.0.1
colorama==0.4.4
commonmark==0.9.1
cryptography==35.0.0
dnspython==2.1.0
future==0.18.2
idna==3.2
iniconfig==1.1.1
ipaddress==1.0.23
isort==5.10.1
Jinja2==3.0.2
junos-eznc==2.6.3
loguru==0.5.3
lxml==4.6.5
MarkupSafe==2.0.1
mypy-extensions==0.4.3
napalm==3.3.1
ncclient==0.6.9
netaddr==0.8.0
netmiko==3.4.0
nornir==3.1.1
nornir-napalm==0.1.2
nornir-netbox==0.3.0
nornir-utils==0.1.2
ntc-templates==2.3.2
packaging==21.3
paramiko==2.7.2
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.4.0
pluggy==1.0.0
py==1.11.0
pycparser==2.20
pyeapi==0.8.4
pyflakes==2.4.0
Pygments==2.10.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyserial==3.5
pytest==6.2.5
python-dotenv==0.19.2
python-netbox==0.0.21
PyYAML==5.4.1
regex==2021.9.30
requests==2.26.0
rich==10.12.0
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.6
scp==0.14.1
six==1.16.1
tenacity==8.0.1
textfsm==1.1.2
toml==0.10.2
tomli==1.2.1
transitions==0.8.10
typing-extensions==3.10.0.2
urllib3==1.26.7
yamlordereddictloader==0.4.0
